requirements.txt
asgiref==3.11.1
click==8.3.1
croniter==6.0.0
Django==6.0.2
django-redis==6.0.0
django-rq==3.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.2
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
rq==2.6.1
six==1.17.0
sqlparse==0.5.5
whitenoise==6.11.0
